onLongPressでうごくように
